Add switch between macro and text input
