+++ profile_uart/src/main/java/no/nordicsemi/android/uart/view/UARTState.kt
@@ -10,7 +10,8 @@ internal data class UARTViewState(
     val selectedConfigurationName: String? = null,
     val isConfigurationEdited: Boolean = false,
     val configurations: List<UARTConfiguration> = emptyList(),
-    val uartManagerState: HTSManagerState = NoDeviceState
+    val uartManagerState: HTSManagerState = NoDeviceState,
+    val isInputVisible: Boolean = true
 ) {
     val showEditDialog: Boolean = editedPosition != null
