+++ profile_uart/src/main/java/no/nordicsemi/android/uart/view/UARTViewEvent.kt
@@ -23,3 +23,5 @@ internal object DisconnectEvent : UARTViewEvent()
 
 internal object NavigateUp : UARTViewEvent()
 internal object OpenLogger : UARTViewEvent()
+
+internal object MacroInputSwitchClick : UARTViewEvent()
